Clean up interruption functions and gpios
the current interrupt functions are a little weird, because
int1 is used sometimes for jack detection but other for bp.
This definitions has been changed and now there are only 2
interruptions functions, only for jack always and the other
for but as well.

INT1 and INT2 flags have been removed and all the logic is
only using MICBIAS or CODEC button press flags.

Signed-off-by: Francisco Mendez <francisco.mendez@intel.com>