Soundwire: register intel sdw master driver with driver_register()
dev->power.no_pm_callbacks flag is used to check if there is pm ops
for the device. Pm callback function will not be invoked if there is
no pm ops. Currently the flag is true for sdw-master devices which
is incorrect since we do have pm ops in intel_sdw_driver. As a result,
intel_suspend() which is the suspend callback function of intel-sdw
driver may not be invoked in system suspend.
With driver_register() and setting md->dev.driver before
device_register(), dev->power.no_pm_callbacks flag will be set to the
correct value (false) by device_pm_check_callbacks() which will be
triggered by device_register().

Signed-off-by: Bard Liao <yung-chuan.liao@linux.intel.com>