ASoC: Intel: sdw_rt711_rt1308_rt715: add multi-cpu DAI support
In original code, one BE DAI is generated for each sdw link with
only one CPU DAI. But on some platforms, multiple speakers on
different sdw links are designed to support one PCM stream. In
this case there are multiple CPU DAI for each sdw link in one
BE DAI.

This driver processes each sdw link info and creates one BE DAI.
It is not fit for above multi-CPU-DAI case. This patch introduce
another sdw dai link, which merges multi-codec on different sdw
link into one sdw dai so that it is easy to gerenate BE DAI with
enough information.

Signed-off-by: randerwang <rander.wang@linux.intel.com>