ASoC: soc-acpi: add flags for generic cards
The existing approach used so far by ASoC driver for Intel platforms
based on a 'monolithic' audio card. That proves to be difficult to
maintain and require the addition of the same definitions in multiple
topologies.

A more elegant solution would consist in a split by 'generic' cards,
where independent functionality is handled separately. That could
enable the use of a separate card for Display Audio (HDMI/DisplayPort)
which is typically generic. The same goes for digital microphones in
simple configurations, or Bluetooth Audio.

The problem is however that interfaces may not be completely
independent. If an amplifier is connected e.g. with an SSP and the
microphones with a separate DMIC interface, the use of acoustic echo
cancellation adds a dependency between interfaces, which results in
topologies and cards depending on external sources/sinks.

This patch suggests new flags to explicitly describe whether generic
cards can be used in an independent manner on each hardware board. In
case there is any sort of coupling, the expectation is that the 'main'
topology for this hardware describes the dependencies and can be
tested in a stand-alone manner.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>