ASoC: SOF: Intel: hda-mlink: introduce helpers for 'extended links' PM
Add helpers to program SPA/CPA bits, using a mutex to access the
shared LCTL register if required, and special-case them for the new
'extended' links.

It is arguably strange to special-case when all links are managed with
the same LCTLx.SPA bits. However the devil is in the details:

Legacy HDaudio multi-links are powered-up when exiting reset.

Alternate links for SoundWire/DMIC/SSP need to be explicitly
powered-up before accessing the SHIM/IP/VS_SHIM space for each
sublink.

In additional, DMIC/SSP/SoundWire are all different cases with a
different device/dai/hlink relationship.

The SSP has a 1:1 mapping between sublink and DAI, and SoundWire will
handle power management with the auxiliary device resume/suspend
routine.

The use of a ref_count is a special-case for the DMIC interface, where
multiple DAIs are exposed for the same link. ref-counting is necessary
for this case only if both DAIs are used concurrently.

And last the SoundWire Intel require power-up/down and bank switch to
be handled with a lock already taken, so the 'eml_lock' is made
optional with the _unlocked versions of the helpers.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>