ASoC: soc-pcm: filter out invalid BEs with NULL substream
A BE is used without an allocated substream is either an error case or
a configuration where the BE is notional, e.g. for DSP-generated echo
reference.

The existing code has multiple filters in startup/trigger, this patch
suggesting plugging the remaining holes in DAI hw_params and stop
cases.

Link: https://github.com/thesofproject/linux/issues/4083
Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>