ASoC: SOF: Intel: hda-stream: fix FIFO size for capture
For some reason the code contributed in 2019 sets the FIFO size to
zero. There's just no rationale for this, the hdac_stream code does
not have any capture-specific behavior and the hardware documentation
does mention that even on capture the driver must write to the SDxFMR
register to kick off the FIFO size calculation, and read-back to find
out the HW allocated FIFO size. Knowing the FIFO size is require to
understand the L1 entry/exit threshold settings in EM registers.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>