cadence: use IDA-based allocation for PDIs
For the longest time, we've assumed a 1:1:1 mapping between PDI, port
and dai->id.

If we want to support high-resolution streams (192kHz, 8ch, 24 bits),
this simplification needs to be revisited: we will need two lanes per
stream, and hence one logical DAI will rely on two PDIs/ports.

To start preparing for this change, this patch changes the PDI
allocation and relies on an IDA mechanism. The Cadence IP supports 3
types of PDIs, so we have to deal with 3 separates IDAs, however in
practice all existing SOCs only rely on Bi-directional PDIs.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>