soundwire: bus: start bus at default frequency
When platform firmware exposes multiple supported bus frequencies, the
existing SoundWire support selects the maximum frequency. This is not
aligned with the SoundWire 1.2 directions: the MIPI recommendation is
to start at a 'safe' speed, compatible with the default frame rate and
shape, and only increase the clock when vendor and codec PHY
parameters are updated.

The clock changes are not supported for now, so in practice this patch
has the effect of discarding frequencies different to the implicit
default.

On Intel platforms with a 'standard' DSDT, this forces the use of
4.8MHz. On QCOM platforms this patch has no effect since only one
frequency is supported.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>