regmap: add async_raw_write_lock/unlock bus callback
The current implementation of regmap_raw_write_async() assumes the
async capability is always available. That assumption is problematic
for SoundWire if we want to use this async support for BTP/BRA: this
async support may not be implemented in all versions of SoundWire or
may be dependent on other audio workloads.

This patch suggests a lock/unlock mechanism around the bus async
write. No functionality change for existing SPI-based platforms.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>