regmap-sdw: add async_lock/unlock/write indirections
The regmap_raw_write_async() will use these indirections. The
functionality is completely hardware-dependent and implemented in
manager ops.

Note that the async_write() does not wait, the wait must be
implemented by the Manager driver which shall use the notification
helper to signal the regmap layer all transactions are complete.

Signed-off-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>